# swagger.yaml
openapi: 3.0.0
info:
  title: DeepThought_NudgeAPI by Manish Kumar Panda
  contact: {}
  version: '1.0'
servers:
# - url: http://localhost:5000/api/v3/app
- url: https://deepthought-nudge-api.onrender.com/
  variables: {}
paths:
...
  /events:
    get:
      tags:
      - Events
      summary: Get All Events
      operationId: GetAllEvents
      parameters:
      - name: id
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: type
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: string
      - name: limit
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      - name: page
        in: query
        description: ''
        required: true
        style: form
        explode: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
    post:
      tags:
      - Events
      summary: Create Event
      operationId: CreateEvent
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateEventRequest'
              - example:
                  category: 6478391d3e6c3f6a30160321
                  subcategory: 647839533e6c3f6a30160325
                  name: Python For Beginners
                  user: 64777fd68377a1a50d06f9a3
                  moderator: 647839b63e6c3f6a3016032b
                  rigor_rank: 9
                  attendees:
                  - 647780518377a1a50d06f9a8
                  - 64777fd68377a1a50d06f9a3
                  imageUrl: dfsdsfsfds
                  description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
                  scheduled:
                  - on: 2023-10-08
                    from: 2023-10-08 9:00:00
                    to: 2023-10-08 13:00:00
                  tagline: Python is future. Learn it now.
            example:
              category: 6478391d3e6c3f6a30160321
              subcategory: 647839533e6c3f6a30160325
              name: Python For Beginners
              user: 64777fd68377a1a50d06f9a3
              moderator: 647839b63e6c3f6a3016032b
              rigor_rank: 9
              attendees:
              - 647780518377a1a50d06f9a8
              - 64777fd68377a1a50d06f9a3
              imageUrl: dfsdsfsfds
              description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
              scheduled:
              - on: 2023-10-08
                from: 2023-10-08 9:00:00
                to: 2023-10-08 13:00:00
              tagline: Python is future. Learn it now.
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /events/{id}:
    put:
      tags:
      - Events
      summary: Update Event
      description: try 64783a7e3e6c3f6a30160330
      operationId: UpdateEvent
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateEventRequest'
              - example:
                  category: 6478391d3e6c3f6a30160321
                  subcategory: 647839533e6c3f6a30160325
                  name: Python For Beginners
                  user: 64777fd68377a1a50d06f9a3
                  moderator: 647839b63e6c3f6a3016032b
                  rigor_rank: 9
                  attendees:
                  - 647780518377a1a50d06f9a8
                  - 64777fd68377a1a50d06f9a3
                  imageUrl: ''
                  description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
                  scheduled:
                  - on: 2023-10-08
                    from: 2023-10-08 9:00:00
                    to: 2023-10-08 13:00:00
                  tagline: Python is future. Learn it now.
            example:
              category: 6478391d3e6c3f6a30160321
              subcategory: 647839533e6c3f6a30160325
              name: Python For Beginners
              user: 64777fd68377a1a50d06f9a3
              moderator: 647839b63e6c3f6a3016032b
              rigor_rank: 9
              attendees:
              - 647780518377a1a50d06f9a8
              - 64777fd68377a1a50d06f9a3
              imageUrl: ''
              description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
              scheduled:
              - on: 2023-10-08
                from: 2023-10-08 9:00:00
                to: 2023-10-08 13:00:00
              tagline: Python is future. Learn it now.
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Events
      summary: Delete Event
      operationId: DeleteEvent
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /images/{id}/upload-image:
    put:
      tags:
      - Events
      summary: Update Event Image
      description: 64783a7e3e6c3f6a30160330
      operationId: UpdateEventImage
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /images/upload-image:
    post:
      tags:
      - Events
      summary: Add Event Image
      operationId: AddEventImage
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /nudge/image/{id}/upload-image:
    put:
      tags:
      - Nudge
      summary: Update Uploaded Nudge Image Cover
      operationId: UpdateUploadedNudgeImageCover
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - image
              type: object
              properties:
                image:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register
      operationId: Register
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/RegisterRequest'
              - example:
                  name: Super User
                  email: superuser123@gmail.com
                  password: wrongpassword
            example:
              name: Super User
              email: superuser123@gmail.com
              password: wrongpassword
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /auth/login:
    post:
      tags:
      - Auth
      summary: Login
      operationId: Login
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/LoginRequest'
              - example:
                  email: superuser123@gmail.com
                  password: please-enter-password
            example:
              email: superuser123@gmail.com
              password: please-enter-password
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
components:
  schemas:
...
    CreateEventRequest:
      title: CreateEventRequest
      required:
      - category
      - subcategory
      - name
      - user
      - moderator
      - rigor_rank
      - attendees
      - imageUrl
      - description
      - scheduled
      - tagline
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
        name:
          type: string
        user:
          type: string
        moderator:
          type: string
        rigor_rank:
          type: integer
          format: int32
        attendees:
          type: array
          items:
            type: string
          description: ''
        imageUrl:
          type: string
        description:
          type: string
        scheduled:
          type: array
          items:
            $ref: '#/components/schemas/Scheduled'
          description: ''
        tagline:
          type: string
      example:
        category: 6478391d3e6c3f6a30160321
        subcategory: 647839533e6c3f6a30160325
        name: Python For Beginners
        user: 64777fd68377a1a50d06f9a3
        moderator: 647839b63e6c3f6a3016032b
        rigor_rank: 9
        attendees:
        - 647780518377a1a50d06f9a8
        - 64777fd68377a1a50d06f9a3
        imageUrl: dfsdsfsfds
        description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
        scheduled:
        - on: 2023-10-08
          from: 2023-10-08 9:00:00
          to: 2023-10-08 13:00:00
        tagline: Python is future. Learn it now.
    Scheduled:
      title: Scheduled
      required:
      - on
      - from
      - to
      type: object
      properties:
        on:
          type: string
        from:
          type: string
        to:
          type: string
      example:
        on: 2023-10-08
        from: 2023-10-08 9:00:00
        to: 2023-10-08 13:00:00
    UpdateEventRequest:
      title: UpdateEventRequest
      required:
      - category
      - subcategory
      - name
      - user
      - moderator
      - rigor_rank
      - attendees
      - imageUrl
      - description
      - scheduled
      - tagline
      type: object
      properties:
        category:
          type: string
        subcategory:
          type: string
        name:
          type: string
        user:
          type: string
        moderator:
          type: string
        rigor_rank:
          type: integer
          format: int32
        attendees:
          type: array
          items:
            type: string
          description: ''
        imageUrl:
          type: string
        description:
          type: string
        scheduled:
          type: array
          items:
            $ref: '#/components/schemas/Scheduled'
          description: ''
        tagline:
          type: string
      example:
        category: 6478391d3e6c3f6a30160321
        subcategory: 647839533e6c3f6a30160325
        name: Python For Beginners
        user: 64777fd68377a1a50d06f9a3
        moderator: 647839b63e6c3f6a3016032b
        rigor_rank: 9
        attendees:
        - 647780518377a1a50d06f9a8
        - 64777fd68377a1a50d06f9a3
        imageUrl: ''
        description: Code with harry teaches python really well. With practical coding exercises, you can learn and be fluent in all concepts in Python.
        scheduled:
        - on: 2023-10-08
          from: 2023-10-08 9:00:00
          to: 2023-10-08 13:00:00
        tagline: Python is future. Learn it now.
...
    RegisterRequest:
      title: RegisterRequest
      required:
      - name
      - email
      - password
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      example:
        name: Super User
        email: superuser123@gmail.com
        password: wrongpassword
    LoginRequest:
      title: LoginRequest
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      example:
        email: superuser123@gmail.com
        password: please-enter-password
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...